Barcode Scanner SDK for Windows v3.6.1.0 Release

+++ SampleApp_CPP/ConfigurationDlg.cpp
@@ -305,7 +305,7 @@ void CConfigurationDlg::UpdateClaimedStatus(int scnID)
 void CConfigurationDlg::OnChangeProtocol()
 {
 	int index = m_cmbProtocol.GetCurSel();
-	int NewProtocol = m_cmbProtocol.GetItemData(index);
+	int NewProtocol = (int) m_cmbProtocol.GetItemData(index);
 	int &CurrentProtocol = QueryScannerProtocol();
 	if(NewProtocol != CurrentProtocol)
 	{
